+++ search/ff_heuristic.h
@@ -148,7 +148,7 @@ class FFHeuristic : public Heuristic {
     void collect_relaxed_plan(Proposition *goal, RelaxedPlan &relaxed_plan, const State &state);
 
     int compute_hsp_add_heuristic();
-    int compute_hsp_max_heuristic();
+    float compute_hsp_max_heuristic();
     int compute_ff_heuristic(const State &state);
 
     void collect_ha(Proposition *goal, RelaxedPlan &relaxed_plan, const State &state);
@@ -158,7 +158,7 @@ class FFHeuristic : public Heuristic {
 protected:
     virtual int compute_heuristic(const State &state);
 public:
-    int get_lower_bound(const State &state);
+    float get_lower_bound(const State &state);
     void set_additional_goals(const std::vector<std::pair<int, int> >& goals);
     void set_recompute_heuristic() {heuristic_recomputation_needed = true;}
     void compute_reachability_with_excludes(std::vector<std::vector<int> >& lvl_var,
